% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installMecabKoForWindows.R
\name{installMecabKoDicForWindows}
\alias{installMecabKoDicForWindows}
\title{install mecab-ko-dic for windows}
\usage{
installMecabKoDicForWindows(path = "c://mecab", cleanup = TRUE)
}
\arguments{
\item{path}{the path to download a file and unzip. The default is "c:/mecab" and Default path is strongly recommended.}

\item{cleanup}{remove download file after unzip. The default is TRUE but you can remain file cleanup = FALSE.}
}
\value{
None.
}
\description{
install mecab-ko-dic for windows
}
\examples{
\dontrun{
installMecabKoDicForWindows()
installMecabKoDicForWindows(path = "C://mecab")
}

}
