% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installMecabKoForWindows.R
\name{installMecabKoForWindows}
\alias{installMecabKoForWindows}
\title{install mecab-ko for windows}
\usage{
installMecabKoForWindows(path = "c:\\\\mecab", release = "auto",
  cleanup = TRUE)
}
\arguments{
\item{path}{The path to download a file and unzip. The default is "c:/mecab" and Default path is strongly recommended.}

\item{release}{"x86" or "x84". The default is "auto" but you can select if you already know.}

\item{cleanup}{Remove download file after unzip. The default is TRUE but you can remain file cleanup = FALSE.}
}
\value{
None.
}
\description{
install mecab-ko for windows
}
\examples{
\dontrun{
installMecabKoForWindows()
installMecabKoForWindows(path = "C://mecab")
}

}
