% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.r
\name{pos}
\alias{pos}
\title{part-of-speech tagger}
\usage{
pos(sentence, dict = "")
}
\arguments{
\item{sentence}{A character vector.}

\item{dict}{A location of user-specific MeCab dictionary. The default value is "".}
}
\value{
A string vector of POS tagged morpheme will be returned in conjoined character
 vecter form. Element name of the list are original phrases
}
\description{
\code{pos} returns part-of-speech (POS) tagged morpheme of the sentence.
}
\details{
This is a basic function for MeCab part-of-speech tagger.
}
\examples{
\dontrun{
pos("")
}

}
