% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppMeCab-package.r
\docType{package}
\name{RcppMeCab}
\alias{RcppMeCab}
\alias{RcppMeCab-package}
\title{RcppMeCab: Rcpp Wrapper for MeCab}
\description{
R package based on \code{Rcpp} for \code{MeCab}: Yet Another Part-of-Speech and
Morphological Analyzer (\url{http://taku910.github.io/mecab/}). The purpose of
this package is providing a seemless developing and analyzing environment for
CJK users.
}
\details{
This package utilizes \code{MeCab} C API and \code{Rcpp} codes.
}
\references{
\itemize{
\item{\href{http://taku910.github.io/mecab/}{MeCab}}
\item{\href{http://www.rcpp.org}{Rcpp: Seamless R and C++ Integration}}
\item{\href{http://eunjeon.blogspot.com}{Eunjeon project}}
}
}
\author{
Junhewk Kim
}
\keyword{Chinese}
\keyword{Japanese}
\keyword{Korean}
\keyword{MeCab}
\keyword{morpheme}
\keyword{nlp}
\keyword{part-of-speech}
